requirements.txt
opencv-python==4.8.1.78
ultralytics==8.0.196
torch>=1.13.0
torchvision>=0.14.0
numpy>=1.21.0
Pillow>=9.0.0
matplotlib>=3.5.0
transformers>=4.21.0
sentence-transformers>=2.2.0
scikit-learn>=1.1.0
colorama>=0.4.5
# Optional CLIP backends:
# openai-clip
# open-clip-torch 